Use new nodespec_to_hostlist() to delete nodes
Bug 13679

+++ src/slurmctld/node_mgr.c
@@ -4801,8 +4801,7 @@ static int _delete_node(char *name)
 extern int delete_nodes(char *names, char **err_msg)
 {
 	char *node_name;
-	hostset_t to_delete;
-	hostlist_iterator_t to_delete_itr;
+	hostlist_t to_delete;
 	bool one_success = false;
 	int ret_rc = SLURM_SUCCESS;
 	hostlist_t error_hostlist = NULL;
@@ -4820,9 +4819,18 @@ extern int delete_nodes(char *names, char **err_msg)
 
 	lock_slurmctld(write_lock);
 
-	to_delete = hostset_create(names);
-	to_delete_itr = hostset_iterator_create(to_delete);
-	while ((node_name = hostlist_next(to_delete_itr))) {
+	if (!(to_delete = nodespec_to_hostlist(names, NULL))) {
+		ret_rc = ESLURM_INVALID_NODE_NAME;
+		goto cleanup;
+	}
+	if (!hostlist_count(to_delete)) {
+		info("%s: expansion of node specification '%s' resulted in zero nodes",
+		     __func__, names);
+		ret_rc = ESLURM_INVALID_NODE_NAME;
+		goto cleanup;
+	}
+
+	while ((node_name = hostlist_shift(to_delete))) {
 		int rc;
 		if ((rc = _delete_node(node_name))) {
 			error("failed to delete node '%s'", node_name);
@@ -4846,10 +4854,10 @@ extern int delete_nodes(char *names, char **err_msg)
 		xfree(nodes);
 	}
 
+cleanup:
 	unlock_slurmctld(write_lock);
 
-	hostlist_iterator_destroy(to_delete_itr);
-	hostset_destroy(to_delete);
+	FREE_NULL_HOSTLIST(to_delete);
 
 	return ret_rc;
 }
